Revert "AnalogWrite logic flipped when I added a pnp transistor. This compensates for that."
This reverts commit bea57b697bfc0ff08cecdeb2bc7a4eb39a48a48f.

I just read that a PNP transistor likely isn't what I want as I may
have a much higher voltage source powering lights an such than the
arduino pinout.

From: https://learn.sparkfun.com/tutorials/transistors

Similar to the NPN circuit, the base is our input, and the emitter is tied to a
constant voltage. This time however, the emitter is tied high, and the load is
connected to the transistor on the ground side.

This circuit works just as well as the NPN-based switch, but there’s one huge
difference: to turn the load “on” the base must be low. This can cause
complications, especially if the load’s high voltage (VCC in this picture) is
higher than our control input’s high voltage. For example, this circuit
wouldn’t work if you were trying to use a 5V-operating Arduino to switch on a
12V motor. In that case it’d be impossible to turn the switch off because VB
would always be less than VE.